♻️  router coding_gate + obsidian_gate 추출 (P0-P step 7+8)
변경 이유
- "코딩 권한 제안 / 수정 승인" 흐름과 "저장 승인 / 이대로 저장"
  흐름이 둘 다 routing preflight 보다 먼저 fire 하는 gate 라 ``_legacy``
  안에서 ~600 줄을 차지하면서 다른 책임과 entangled 되어 있었다.
  각자 단독 모듈로 분리해 향후 새 gate 추가 시 변경 표면을 한 파일로.

주요 변경 사항
- 새 파일 ``coding_gate.py`` (약 300 줄):
  - ``_find_session_with_pending_coding_proposal``
  - ``_find_latest_open_session``
  - ``_run_coding_authorization_gate``
  - phrase_detect / coding.authorization / coding.job / proposal
    serialization helper 의존 명시 import.
- 새 파일 ``obsidian_gate.py`` (약 310 줄):
  - ``_can_save_to_obsidian``
  - ``_run_obsidian_approval_gate``
  - ``_run_obsidian_preview_branch``
  - ``_find_session_with_pending_proposal``
  - obsidian.approval / lifecycle.resolver 의존 명시 import
    (``build_save_proposal``, ``execute_pending_proposal``,
    ``store_pending_proposal``, ``ObsidianApprovalError``,
    ``is_obsidian_approval``/``_save_request``,
    ``extract_explicit_session_id`` 등).
- ``_legacy.py`` 의 해당 7 함수 (517 줄) 제거 → re-export 만 남김.
- ``__init__.py`` facade 가 두 gate 직접 import.
- session_persistence (step 6) 의 helper 들을 ``_legacy.py`` 내부 call
  site 에서 다시 쓸 수 있도록 re-import 블록도 추가.

비고
- 4672 PASS + 1 skipped 회귀 유지.
- 각 gate 의 nested import (~8-10 종) 모두 module top 으로 끌어올려
  반복 lazy-import overhead 제거.
