+++ src/yule_orchestrator/discord/engineering_channel_router/_legacy.py
@@ -905,115 +905,10 @@ from ..engineering.phrase_detect import (
 )
 
 
-def _find_session_with_pending_coding_proposal(
-    *,
-    message: Any,
-    list_sessions_fn: Callable[..., Sequence[Any]],
-) -> Optional[Any]:
-    """Pick the session whose ``extra['coding_proposal']`` should pair
-    with this approval phrase. Mirrors ``_find_session_with_pending_proposal``
-    but reads the coding key instead of the obsidian key."""
-
-    try:
-        try:
-            sessions = list_sessions_fn(limit=50)
-        except TypeError:
-            sessions = list_sessions_fn()
-    except Exception:  # noqa: BLE001
-        return None
-    if not sessions:
-        return None
-
-    channel = getattr(message, "channel", None)
-    channel_id = getattr(channel, "id", None)
-    parent_id = getattr(channel, "parent_id", None)
-    if parent_id is None and getattr(channel, "parent", None) is None:
-        thread_id = None
-        scoped_channel_id = channel_id
-    else:
-        thread_id = channel_id
-        scoped_channel_id = parent_id
-    user_id = getattr(getattr(message, "author", None), "id", None)
-
-    candidates = [
-        s
-        for s in sessions
-        if isinstance(getattr(s, "extra", None), Mapping)
-        and dict(getattr(s, "extra")).get("coding_proposal")
-    ]
-    if not candidates:
-        return None
-
-    if thread_id is not None:
-        for session in candidates:
-            if getattr(session, "thread_id", None) == thread_id:
-                return session
-
-    if scoped_channel_id is not None:
-        same_scope = [
-            s
-            for s in candidates
-            if getattr(s, "channel_id", None) == scoped_channel_id
-            and (user_id is None or getattr(s, "user_id", None) == user_id)
-        ]
-        if same_scope:
-            return _most_recent_session(same_scope)
-
-    return _most_recent_session(candidates)
-
-
-def _find_latest_open_session(
-    *,
-    message: Any,
-    list_sessions_fn: Callable[..., Sequence[Any]],
-) -> Optional[Any]:
-    """Pick the session a coding proposal should target when the user
-    didn't reference one explicitly. Same channel/thread > same channel
-    > most recently updated open session."""
-
-    try:
-        try:
-            sessions = list_sessions_fn(limit=50)
-        except TypeError:
-            sessions = list_sessions_fn()
-    except Exception:  # noqa: BLE001
-        return None
-    if not sessions:
-        return None
-
-    open_sessions = [s for s in sessions if not _is_terminal(s)]
-    if not open_sessions:
-        return None
-
-    channel = getattr(message, "channel", None)
-    channel_id = getattr(channel, "id", None)
-    parent_id = getattr(channel, "parent_id", None)
-    if parent_id is None and getattr(channel, "parent", None) is None:
-        thread_id = None
-        scoped_channel_id = channel_id
-    else:
-        thread_id = channel_id
-        scoped_channel_id = parent_id
-
-    if thread_id is not None:
-        for session in open_sessions:
-            if getattr(session, "thread_id", None) == thread_id:
-                return session
-
-    if scoped_channel_id is not None:
-        same_scope = [
-            s
-            for s in open_sessions
-            if getattr(s, "channel_id", None) == scoped_channel_id
-        ]
-        if same_scope:
-            return _most_recent_session(same_scope)
-
-    return _most_recent_session(open_sessions)
-
-
-# P0-P step 6: session.extra mutations + load helpers extracted to .session_persistence.
-from .session_persistence import (  # noqa: E402,F401 — re-export for back-compat
+# P0-P step 6: session.extra mutations + load helpers extracted to
+# .session_persistence. _legacy still calls these — re-import so the
+# local bindings stay resolvable inside the orchestration functions.
+from .session_persistence import (  # noqa: E402,F401 — re-import for local call sites
     _is_terminal,
     _load_session_by_id,
     _most_recent_session,
@@ -1029,6 +924,23 @@ from .session_persistence import (  # noqa: E402,F401 — re-export for back-com
     _record_persistence_failure,
     _work_report_to_dict,
 )
+# P0-P step 7: coding 권한 / 승인 gate extracted to .coding_gate.
+from .coding_gate import (  # noqa: E402,F401 — re-export for back-compat
+    _find_session_with_pending_coding_proposal,
+    _find_latest_open_session,
+    _run_coding_authorization_gate,
+)
+# P0-P step 8: Obsidian 저장 gate extracted to .obsidian_gate.
+from .obsidian_gate import (  # noqa: E402,F401 — re-export for back-compat
+    _can_save_to_obsidian,
+    _run_obsidian_approval_gate,
+    _run_obsidian_preview_branch,
+    _find_session_with_pending_proposal,
+)
+
+
+
+
 
 
 
@@ -1130,410 +1042,14 @@ async def _emit_work_report_preview(
 
 
 
-async def _run_coding_authorization_gate(
-    *,
-    message: Any,
-    prompt_text: str,
-    list_sessions_fn: Callable[..., Sequence[Any]],
-    send_chunks: SendChunksFn,
-) -> Optional[EngineeringRouteResult]:
-    """Two-branch gate.
-
-    1. ``is_coding_proposal_request`` — build a fresh proposal and
-       stash it under ``session.extra['coding_proposal']``, then post
-       the preview. The user follows up with an approval phrase.
-    2. ``is_coding_approval_phrase`` — flip the latest stashed
-       proposal into a ``CodingJob`` (status=ready) and persist under
-       ``session.extra['coding_job']``.
-
-    Returns ``None`` when the message isn't either kind so the caller
-    falls through to the rest of the route.
-    """
-
-    # Hard "no code change" override: if the user explicitly said
-    # "코드 수정 하지 말고 리서치만" the coding gate must not act on
-    # this message even if it also contains a proposal/approval phrase.
-    if _user_explicitly_blocked_coding(prompt_text):
-        return None
-
-    if is_coding_proposal_request(prompt_text):
-        target = _find_latest_open_session(
-            message=message,
-            list_sessions_fn=list_sessions_fn,
-        )
-        if target is None:
-            await send_chunks(
-                message.channel,
-                (
-                    "현재 채널에 매칭되는 열린 engineering-agent 세션이 보이지 않아요.\n"
-                    "먼저 작업을 접수해서 세션을 만들고 다시 `코딩 권한 제안`이라고 답해 주세요."
-                ),
-            )
-            return EngineeringRouteResult(handled=True)
-
-        proposal = recommend_authorization(
-            user_request=getattr(target, "prompt", "") or "",
-            session_id=getattr(target, "session_id", None),
-        )
-        _persist_coding_proposal(target, proposal)
-        await send_chunks(message.channel, format_authorization_message(proposal))
-        return EngineeringRouteResult(
-            handled=True,
-            session_id=getattr(target, "session_id", None),
-            thread_id=getattr(target, "thread_id", None),
-        )
-
-    if is_coding_approval_phrase(prompt_text):
-        owner = _find_session_with_pending_coding_proposal(
-            message=message,
-            list_sessions_fn=list_sessions_fn,
-        )
-        if owner is None:
-            await send_chunks(
-                message.channel,
-                (
-                    "지금은 대기 중인 코딩 권한 제안이 없어요.\n"
-                    "먼저 `코딩 권한 제안` 이라고 답해서 Tech Lead 추천을 받아 주세요."
-                ),
-            )
-            return EngineeringRouteResult(handled=True)
-
-        extra = dict(getattr(owner, "extra", {}) or {})
-        payload = extra.get("coding_proposal")
-        if not isinstance(payload, Mapping):
-            await send_chunks(
-                message.channel,
-                "대기 중인 코딩 권한 제안 payload를 읽지 못했어요. 다시 `코딩 권한 제안`을 시도해 주세요.",
-            )
-            return EngineeringRouteResult(handled=True)
-
-        from datetime import datetime as _dt
-        from datetime import timezone as _tz
-
-        approved_at = _dt.now(_tz.utc)
-        proposal = _proposal_from_dict(payload)
-        try:
-            job = build_coding_job_from_proposal(
-                proposal,
-                status=STATUS_READY,
-                approved_at=approved_at,
-            )
-        except Exception as exc:  # noqa: BLE001
-            await send_chunks(
-                message.channel,
-                f"⚠️ 코딩 권한 승인 중 오류가 발생했어요: {exc}",
-            )
-            return EngineeringRouteResult(handled=True, error=str(exc))
-
-        _persist_coding_job(owner, job.to_dict())
-
-        thread_label = (
-            f"thread `{job.session_id}`"
-            if job.session_id
-            else "(session id 미기록)"
-        )
-        await send_chunks(
-            message.channel,
-            "\n".join(
-                [
-                    "**[engineering-agent] 코딩 권한 승인 완료**",
-                    "",
-                    f"executor: `{job.executor_role}`",
-                    f"세션: {thread_label}",
-                    f"승인 시각: {approved_at.isoformat()}",
-                    "",
-                    "이제 executor에게 안전한 prompt가 전달될 준비가 됐어요. 실제 코드 변경은 executor가 계획을 보여 드린 뒤에만 진행합니다.",
-                ]
-            ),
-        )
-        return EngineeringRouteResult(
-            handled=True,
-            session_id=getattr(owner, "session_id", None),
-            thread_id=getattr(owner, "thread_id", None),
-        )
-
-    return None
 
 
-# MVP closure refactor — explicit session id regex moved to
-# :mod:`agents.lifecycle.resolver` so router / bot / obsidian gate
-# share one canonical implementation. The router-private alias is
-# kept for backward compat with internal callers (and the runtime
-# preflight ``_explicit_session_id`` substring check).
-from ...agents.lifecycle.resolver import (
-    _EXPLICIT_SESSION_ID_RE as _EXPLICIT_SESSION_ID_RE,
-    extract_explicit_session_id as _extract_session_id_from_router_text,
-)
 
 
-def _can_save_to_obsidian(session: Any) -> tuple[bool, Optional[str]]:
-    """Return (allowed, blocking_reason).
-
-    Phase 4 stab: an Obsidian write must NOT proceed when the
-    lifecycle hasn't actually closed. Reads ``session.extra`` for the
-    Phase 2/3 status keys and refuses if research is empty / forum
-    isn't connected / work_report is not ready/final.
-
-    Returns ``(True, None)`` to allow, ``(False, "<korean reason>")``
-    to block. Test stubs that don't carry rich extras get a generous
-    "missing canonical readiness" reason rather than a hard pass.
-    """
-
-    # Refactor: delegate to the canonical :mod:`agents.lifecycle.status`
-    # helper so the router, work_report builder, and Discord status
-    # diagnostic all share one set of "can we save?" rules. The block
-    # reasons stay identical to keep operator-visible messages stable.
-    from ...agents.lifecycle.status import can_write_obsidian_record
-
-    return can_write_obsidian_record(session)
 
 
-async def _run_obsidian_approval_gate(
-    *,
-    message: Any,
-    prompt_text: str,
-    list_sessions_fn: Callable[..., Sequence[Any]],
-    send_chunks: SendChunksFn,
-    writer_fn: Optional[Callable[..., Any]] = None,
-    env: Optional[Any] = None,
-) -> Optional[EngineeringRouteResult]:
-    """Try to interpret *prompt_text* as an Obsidian save approval.
-
-    Phase 4 stab: explicit "세션 <id> 기준으로 저장해줘" prompts now
-    resolve via the id first (load_session) — channel/thread walks
-    only fire when the user didn't name a session. Final write is
-    blocked when the lifecycle is incomplete (no research_pack,
-    interim/insufficient work_report, missing role coverage).
-
-    Returns a populated :class:`EngineeringRouteResult` when the message
-    was an approval phrase (regardless of whether the write succeeded),
-    or ``None`` to fall through to the runtime preflight + conversation
-    flow. We deliberately keep this branch above the runtime classifier
-    so a bare "저장 승인" never gets promoted to ``new_work_request``.
-    """
-
-    # Phase 4 stab: accept "세션 <id> 기준으로 저장 승인" by stripping
-    # the explicit-id preamble before testing the approval phrase.
-    explicit_id = _extract_session_id_from_router_text(prompt_text)
-    test_text = prompt_text
-    if explicit_id and not is_obsidian_approval(test_text):
-        # Drop the "세션 <id> 기준으로" prefix and re-test so a
-        # session-scoped approval still routes through this gate.
-        stripped = _EXPLICIT_SESSION_ID_RE.sub("", prompt_text).strip()
-        if stripped:
-            for filler in ("기준으로", "기준 으로", "기준에서", "기준"):
-                if stripped.startswith(filler):
-                    stripped = stripped[len(filler):].strip()
-                    break
-            if is_obsidian_approval(stripped):
-                test_text = stripped
-
-    if not is_obsidian_approval(test_text):
-        return None
-
-    candidate: Optional[Any] = None
-    if explicit_id:
-        try:
-            from ...agents.workflow_state import load_session as _load_session
-
-            candidate = _load_session(explicit_id)
-        except Exception:  # noqa: BLE001 - lookup failure falls through
-            candidate = None
-        if candidate is None:
-            await send_chunks(
-                message.channel,
-                (
-                    f"세션 `{explicit_id}` 을 찾지 못했어요.\n"
-                    "session id 가 정확한지 확인하거나, 새 작업이라면 `새 작업으로 진행`이라고 답해 주세요."
-                ),
-            )
-            return EngineeringRouteResult(
-                handled=True,
-                error=f"obsidian approval: explicit session {explicit_id} not found",
-            )
-
-    if candidate is None:
-        candidate = _find_session_with_pending_proposal(
-            message=message,
-            list_sessions_fn=list_sessions_fn,
-        )
-    if candidate is None:
-        await send_chunks(
-            message.channel,
-            (
-                "지금은 대기 중인 Obsidian 저장 제안이 없어요.\n"
-                "먼저 `Obsidian에 정리해줘` 처럼 저장 미리보기를 만들어 주세요."
-            ),
-        )
-        return EngineeringRouteResult(handled=True)
-
-    allowed, block_reason = _can_save_to_obsidian(candidate)
-    if not allowed:
-        await send_chunks(
-            message.channel,
-            (
-                "Obsidian 저장을 진행하지 않았어요.\n"
-                f"차단 사유: {block_reason}\n"
-                "lifecycle 이 완료되면 다시 `저장 승인` 으로 답해 주세요."
-            ),
-        )
-        return EngineeringRouteResult(
-            handled=True,
-            session_id=getattr(candidate, "session_id", None),
-            error=f"obsidian approval blocked: {block_reason}",
-        )
-
-    try:
-        updated, outcome = execute_pending_proposal(
-            candidate,
-            env=env,
-            writer_fn=writer_fn,
-        )
-    except ObsidianApprovalError as exc:
-        await send_chunks(message.channel, f"⚠️ Obsidian 저장 실패: {exc}")
-        return EngineeringRouteResult(handled=True, error=str(exc))
-    except Exception as exc:  # noqa: BLE001 — never let the bot crash on save
-        await send_chunks(
-            message.channel,
-            f"⚠️ Obsidian 저장 중 예상치 못한 오류: {exc}",
-        )
-        return EngineeringRouteResult(handled=True, error=str(exc))
-
-    await send_chunks(message.channel, outcome.message)
-    return EngineeringRouteResult(
-        handled=True,
-        session_id=getattr(updated, "session_id", None),
-    )
 
 
-async def _run_obsidian_preview_branch(
-    *,
-    message: Any,
-    prompt_text: str,
-    decision_payload: Any,
-    list_sessions_fn: Callable[..., Sequence[Any]],
-    send_chunks: SendChunksFn,
-) -> Optional[EngineeringRouteResult]:
-    """Render the preview for an Obsidian save request and store the proposal.
-
-    Called from :func:`_run_runtime_preflight` when the runtime intent
-    classifier identified an Obsidian save request and Recall matched a
-    session. We never call ``thread_continuation_fn`` here — joining the
-    thread isn't the user's goal; they want to see what we'd write before
-    approving it.
-    """
-
-    target_session_id: Optional[str] = None
-    if hasattr(decision_payload, "get"):
-        raw = decision_payload.get("session_id")
-        target_session_id = str(raw) if raw is not None else None
-
-    session = _load_session_by_id(list_sessions_fn, target_session_id)
-    if session is None:
-        await send_chunks(
-            message.channel,
-            (
-                "**[engineering-agent] Obsidian 저장 대상 세션을 찾지 못했어요.**\n"
-                "어떤 세션을 저장할지 `기존 세션 <id>` 처럼 답해 주세요."
-            ),
-        )
-        return EngineeringRouteResult(
-            handled=True,
-            error="obsidian preview: matched session not loadable",
-        )
-
-    try:
-        proposal = build_save_proposal(
-            session,
-            actor_user_id=getattr(getattr(message, "author", None), "id", None),
-        )
-    except ObsidianApprovalError as exc:
-        await send_chunks(message.channel, f"⚠️ {exc}")
-        return EngineeringRouteResult(handled=True, error=str(exc))
-    except Exception as exc:  # noqa: BLE001
-        await send_chunks(
-            message.channel,
-            f"⚠️ Obsidian 미리보기 생성 실패: {exc}",
-        )
-        return EngineeringRouteResult(handled=True, error=str(exc))
-
-    try:
-        store_pending_proposal(session, proposal)
-    except Exception as exc:  # noqa: BLE001 — preview survives even if persist fails
-        await send_chunks(
-            message.channel,
-            f"⚠️ 저장 제안 기록 실패: {exc}\n미리보기는 아래에서 확인하실 수 있어요.",
-        )
-
-    await send_chunks(message.channel, proposal.preview_message)
-    return EngineeringRouteResult(
-        handled=True,
-        session_id=session.session_id,
-        thread_id=getattr(session, "thread_id", None),
-    )
-
-
-def _find_session_with_pending_proposal(
-    *,
-    message: Any,
-    list_sessions_fn: Callable[..., Sequence[Any]],
-) -> Optional[Any]:
-    """Return the session that owns the most relevant pending proposal.
-
-    Match priority:
-      1. Sessions whose ``thread_id`` equals the message's channel id
-         (for thread messages — Discord exposes the thread id as
-         ``channel.id`` and the channel id as ``channel.parent_id``).
-      2. Sessions in the same channel + same author with a proposal.
-      3. Most recently updated session with a proposal.
-
-    Returns ``None`` when no candidate has a pending proposal stashed in
-    ``session.extra``.
-    """
-
-    try:
-        try:
-            sessions = list_sessions_fn(limit=50)
-        except TypeError:
-            sessions = list_sessions_fn()
-    except Exception:  # noqa: BLE001 — recall outage must not crash bot
-        return None
-    if not sessions:
-        return None
-
-    channel = getattr(message, "channel", None)
-    channel_id = getattr(channel, "id", None)
-    parent_id = getattr(channel, "parent_id", None)
-    if parent_id is None and getattr(channel, "parent", None) is None:
-        thread_id = None
-        scoped_channel_id = channel_id
-    else:
-        thread_id = channel_id
-        scoped_channel_id = parent_id
-    user_id = getattr(getattr(message, "author", None), "id", None)
-
-    candidates_with_proposal = [
-        s for s in sessions if get_pending_proposal(s) is not None
-    ]
-    if not candidates_with_proposal:
-        return None
-
-    if thread_id is not None:
-        for session in candidates_with_proposal:
-            if getattr(session, "thread_id", None) == thread_id:
-                return session
-
-    if scoped_channel_id is not None:
-        same_scope = [
-            s
-            for s in candidates_with_proposal
-            if getattr(s, "channel_id", None) == scoped_channel_id
-            and (user_id is None or getattr(s, "user_id", None) == user_id)
-        ]
-        if same_scope:
-            return _most_recent_session(same_scope)
-
-    return _most_recent_session(candidates_with_proposal)
